backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

djoser==2.2.2
djangorestframework-simplejwt==5.3.1

psycopg2-binary==2.9.11

django-redis==5.4.0
redis==5.0.1

Pillow==10.1.0

openai>=1.30.0

torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3